move `health` module inside `system`
