+++ bee-storage/bee-storage-sled/src/storage.rs
@@ -8,8 +8,7 @@ use crate::config::{SledConfig, SledConfigBuilder};
 use bee_storage::{
     access::{Fetch, Insert},
     backend::StorageBackend,
-    health::StorageHealth,
-    system::{StorageVersion, System, SYSTEM_HEALTH_KEY, SYSTEM_VERSION_KEY},
+    system::{StorageHealth, StorageVersion, System, SYSTEM_HEALTH_KEY, SYSTEM_VERSION_KEY},
 };
 
 use async_trait::async_trait;
